Adjust useEffect and implement setScore/highscore
UseEffect has functions (dependencies) inside it so no need in dependencyList. Only setPokemonList if it were to change so pokemonList can be included in dependencyList. The onclick fn is in card.jsx as well as the shuffle. Shuffle changes pokemonList (technically just order), makes it rerender cards if they actually different positions.